tt-5: * moved pages to src
