+++ src/pages/homepage/components/task-item/styles.module.scss
