+++ src/pages/homepage/libs/enums/index.ts
