+++ src/pages/homepage/libs/types/index.ts
